# src/main/java/com/example/store/dto/CreateOrderDTO.java
package com.example.store.dto;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;

import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
public class CreateOrderDTO {
    @NotNull(message = "Customer ID is required") private Long customerId;

    private String description;

    @NotEmpty(message = "An order must contain at least one product ID")
    private List<Long> productIds;
}